[mini-browser] one widget per a URI + aligning behaviour with the preview widget
